quant config
